Fix session coupling in tests
