+++ tests/announcement_test.py
@@ -1,39 +1,13 @@
-import tempfile
-import pytest
 from typing import Generator
 from fastapi.testclient import TestClient
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker, Session
-from main import app
-from database import get_db
+import pytest
 from models.announcement import Announcement
+from tests.test_fixtures import create_and_teardown_tables,client
 
 
-def override_get_db() -> Generator[Session, None, None]:
-    try:
-        db = TestingSessionLocal()
-        yield db
-    finally:
-        db.close()
-
-
-@pytest.fixture(scope="module", autouse=True)
+@pytest.fixture(scope="function", autouse=True)
 def setup_and_teardown_db() -> Generator[TestClient, None, None]:
-    Announcement.metadata.create_all(bind=engine)
-    yield client
-    app.dependency_overrides.clear()
-    Announcement.metadata.drop_all(bind=engine)
-    engine.dispose()
-
-
-temp_db_file = tempfile.NamedTemporaryFile(suffix=".db", delete=False)
-SQLALCHEMY_TEST_DATABASE_URL = f"sqlite:///{temp_db_file.name}"
-engine = create_engine(SQLALCHEMY_TEST_DATABASE_URL)
-
-app.dependency_overrides[get_db] = override_get_db
-TestingSessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
-client = TestClient(app)
-
+    yield from create_and_teardown_tables(Announcement.metadata)
 
 def test_create_get_announcement() -> None:
     create_route = "/announcements/"
